Basic RESTful endpoints to support ReactJS front-end app
The "backstage" namespace represents the controllers that are supporting
the (coming soon) ReactJS app.

For operations that are not very simple, I have created services.  Also
we should use JsonPresenters for anything that is not a simple object.

For testing we should have unit tests (of course) but also request test
to verify the JSON coming from the controllers is correct.