Adding Geocoding for Venues
Eventually we may want to support searching venues by area, city,
or geographic distance (PostGIS).  Adding Geocoding now might be
a bit of over-design, but it is easy to implement.

Also included a gigs.priv_hash column to support a public page that only
the user and associated gig_musicians can see.