Renames JSON 'presenter' with more appropriate and general serializer name
I was rushing through the JSON serializer and had presenters on the mind.
I misnamed things and it has really bothered me.  Phew!

The serializer is also going to be more general because I think we will have
only basic needs for serialization.  This means the JSON and XML serilization
will happen in the same object, until (or if) it becomes too complex.